F.A.C.E (Facial Analysis and Classification Expert) is an Image Age Classifier for Information Security using computer vision and machine learning.
F.A.C.E (Facial Analysis and Classification Expert)

FACE is an Image Age Classifier for Information Security using computer vision and machine learning.
It todays world data is available within a hands reach and hence unnecessary influence to data is a big issue.
Our software scans the person's face and then identifies him/her as an adult or a minor and grants access to only adults.
We have used real time image processing and computer vision along with machine learning to achieve this purpose.